Move PassManager declaration to opt.h
