+++ src/opt.cpp
@@ -96,138 +96,6 @@
 #endif
 
 using namespace ispc;
-
-///////////////////////////////////////////////////////////////////////////
-// This is a wrap over class llvm::ModulePassManager. This duplicates PassManager function run()
-//   and adds several add functions with some checks and debug passes.
-//   This wrap can control:
-//   - If we want to switch off optimization with given number.
-//   - If we want to dump LLVM IR after optimization with given number.
-//   - If we want to generate LLVM IR debug for gdb after optimization with given number.
-class DebugModulePassManager {
-  public:
-    DebugModulePassManager(llvm::Module &M, int optLevel) : m_passNumber(0), m_optLevel(optLevel) {
-        m = &M;
-        llvm::Triple targetTriple = llvm::Triple(m->getTargetTriple());
-        llvm::TargetLibraryInfoImpl targetLibraryInfo(targetTriple);
-        targetMachine = g->target->GetTargetMachine();
-
-        // We have to register an llvm::OptNoneInstrumentation with a llvm::PassInstrumentationCallbacks,
-        // which is then registered in the llvm::PassBuilder constructor.
-        // This ensures that any function with optnone will not be optimized.
-        OptNoneInst.registerCallbacks(PIC);
-
-        if (g->debugPMTimeTrace) {
-            // Enable time traces for optimization passes.
-            TimePasses.registerCallbacks(PIC);
-        }
-        // Create the new pass manager builder using our target machine.
-#if ISPC_LLVM_VERSION >= ISPC_LLVM_16_0
-        pb = llvm::PassBuilder(targetMachine, llvm::PipelineTuningOptions(), std::nullopt, &PIC);
-#else
-        pb = llvm::PassBuilder(targetMachine, llvm::PipelineTuningOptions(), llvm::None, &PIC);
-#endif
-
-        // Register all the basic analyses with the managers.
-        pb.registerModuleAnalyses(mam);
-        pb.registerCGSCCAnalyses(cgam);
-        pb.registerFunctionAnalyses(fam);
-        pb.registerLoopAnalyses(lam);
-        pb.crossRegisterProxies(lam, fam, cgam, mam);
-
-#if ISPC_LLVM_VERSION >= ISPC_LLVM_17_0
-        SI.registerCallbacks(PIC, &mam);
-#else
-        SI.registerCallbacks(PIC, &fam);
-#endif
-
-        // Register all the analysis passes
-        fam.registerPass([&] { return targetMachine->getTargetIRAnalysis(); });
-        fam.registerPass([&] { return llvm::TargetLibraryAnalysis(targetLibraryInfo); });
-
-        // Add alias analysis for more aggressive optimizations
-        if (m_optLevel != 0) {
-            llvm::AAManager aam;
-            // The order in which these are registered determines their priority when
-            // being queried.
-
-            // First we register the basic alias analysis that provides the majority of
-            // per-function local AA logic. This is a stateless, on-demand local set of
-            // AA techniques.
-            aam.registerFunctionAnalysis<llvm::BasicAA>();
-
-            // Next we query fast, specialized alias analyses that wrap IR-embedded
-            // information about aliasing.
-            aam.registerFunctionAnalysis<llvm::ScopedNoAliasAA>();
-            aam.registerFunctionAnalysis<llvm::TypeBasedAA>();
-
-            // Add support for querying global aliasing information when available.
-            // Because the `AAManager` is a function analysis and `GlobalsAA` is a module
-            // analysis, all that the `AAManager` can do is query for any *cached*
-            // results from `GlobalsAA` through a readonly proxy.
-            // aam.registerModuleAnalysis<llvm::GlobalsAA>();
-
-            // Add target-specific alias analyses.
-            if (targetMachine) {
-                targetMachine->registerDefaultAliasAnalyses(aam);
-            }
-            fam.registerPass([aam] { return std::move(aam); });
-        }
-    }
-
-    llvm::PreservedAnalyses run() { return mpm.run(*m, mam); }
-    enum Passes { Module, Function, Loop };
-    template <typename T> void addModulePass(T &&P, int stage = -1);
-    template <typename T> void addPostOrderCGSCCPass(T &&P, int stage = -1);
-    template <typename T> void addFunctionPass(T &&P, int stage = -1);
-    template <typename T> void addLoopPass(T &&P, int stage = -1);
-    // Start a new group of function passes
-    void initFunctionPassManager();
-    // Add function passes to the ModulePassManager
-    void commitFunctionToModulePassManager();
-    // Start a new group of loop passes
-    void initLoopPassManager();
-    // Add loop passes to the FunctionPassManager
-    void commitLoopToFunctionPassManager();
-
-    void setMemorySSA(bool v) { m_memorySSA = v; }
-    void setBlocksFreq(bool v) { m_blocksFreq = v; }
-
-  private:
-    llvm::TargetMachine *targetMachine;
-    llvm::PassBuilder pb;
-    llvm::LoopAnalysisManager lam;
-    llvm::FunctionAnalysisManager fam;
-    llvm::CGSCCAnalysisManager cgam;
-    llvm::ModuleAnalysisManager mam;
-    llvm::ModulePassManager mpm;
-    llvm::PassInstrumentationCallbacks PIC;
-    llvm::PrintPassOptions PrintPassOpts{/*Verbose*/ true, /*SkipAnalyses*/ true, /*Indent*/ true};
-#if ISPC_LLVM_VERSION >= ISPC_LLVM_16_0
-    llvm::StandardInstrumentations SI{*g->ctx, /*DebugLogging*/ g->debugPM, /*VerifyEach*/ false, PrintPassOpts};
-#else
-    llvm::StandardInstrumentations SI{/*DebugLogging*/ g->debugPM, /*VerifyEach*/ false, PrintPassOpts};
-#endif
-    llvm::OptNoneInstrumentation OptNoneInst{/*DebugLogging*/ false};
-    llvm::TimePassesHandler TimePasses{true};
-
-    std::vector<std::unique_ptr<llvm::raw_fd_ostream>> outputDebugDumps;
-    std::vector<std::unique_ptr<llvm::FunctionPassManager>> fpmVec;
-    std::vector<std::unique_ptr<llvm::LoopPassManager>> lpmVec;
-
-    llvm::Module *m;
-
-    bool m_isFPMOpen{false};
-    bool m_isLPMOpen{false};
-    bool m_memorySSA{false};
-    bool m_blocksFreq{false};
-    int m_passNumber;
-    int m_optLevel;
-
-    // Add pass to pass manager and print IR if needed
-    void addPassAndDebugPrint(std::string name, DebugModulePassManager::Passes kind);
-};
-
 // Strips all non-alphanumeric characters from given string.
 static std::string lSanitize(std::string in) {
     llvm::Regex r("[^[:alnum:]]");
@@ -259,49 +127,80 @@ static std::string getDumpFilePath(std::string className, int pnum) {
     return pathDirFile;
 }
 
-template <typename T> void DebugModulePassManager::addModulePass(T &&P, int stage) {
-    Assert(!m_isFPMOpen && "FunctionPassManager must be committed before adding module passes.");
-    Assert(!m_isLPMOpen && "LoopPassManager must be committed before adding module passes.");
-    // taking number of optimization
-    m_passNumber = (stage == -1) ? (m_passNumber + 1) : stage;
-    if (g->off_stages.find(m_passNumber) == g->off_stages.end()) {
-        mpm.addPass(std::move(P));
-        addPassAndDebugPrint(T::name().str(), DebugModulePassManager::Passes::Module);
+DebugModulePassManager::DebugModulePassManager(llvm::Module &M, int optLevel) : m_passNumber(0), m_optLevel(optLevel) {
+    m = &M;
+    llvm::Triple targetTriple = llvm::Triple(m->getTargetTriple());
+    llvm::TargetLibraryInfoImpl targetLibraryInfo(targetTriple);
+    targetMachine = g->target->GetTargetMachine();
+
+    // We have to register an llvm::OptNoneInstrumentation with a llvm::PassInstrumentationCallbacks,
+    // which is then registered in the llvm::PassBuilder constructor.
+    // This ensures that any function with optnone will not be optimized.
+    OptNoneInst.registerCallbacks(PIC);
+
+    if (g->debugPMTimeTrace) {
+        // Enable time traces for optimization passes.
+        TimePasses.registerCallbacks(PIC);
+    }
+    // Create the new pass manager builder using our target machine.
+#if ISPC_LLVM_VERSION >= ISPC_LLVM_16_0
+    pb = llvm::PassBuilder(targetMachine, llvm::PipelineTuningOptions(), std::nullopt, &PIC);
+#else
+    pb = llvm::PassBuilder(targetMachine, llvm::PipelineTuningOptions(), llvm::None, &PIC);
+#endif
+
+    // Register all the basic analyses with the managers.
+    pb.registerModuleAnalyses(mam);
+    pb.registerCGSCCAnalyses(cgam);
+    pb.registerFunctionAnalyses(fam);
+    pb.registerLoopAnalyses(lam);
+    pb.crossRegisterProxies(lam, fam, cgam, mam);
+
+#if ISPC_LLVM_VERSION >= ISPC_LLVM_17_0
+    SI.registerCallbacks(PIC, &mam);
+#else
+    SI.registerCallbacks(PIC, &fam);
+#endif
+
+    // Register all the analysis passes
+    fam.registerPass([&] { return targetMachine->getTargetIRAnalysis(); });
+    fam.registerPass([&] { return llvm::TargetLibraryAnalysis(targetLibraryInfo); });
+
+    // Add alias analysis for more aggressive optimizations
+    if (m_optLevel != 0) {
+        llvm::AAManager aam;
+        // The order in which these are registered determines their priority when
+        // being queried.
+
+        // First we register the basic alias analysis that provides the majority of
+        // per-function local AA logic. This is a stateless, on-demand local set of
+        // AA techniques.
+        aam.registerFunctionAnalysis<llvm::BasicAA>();
+
+        // Next we query fast, specialized alias analyses that wrap IR-embedded
+        // information about aliasing.
+        aam.registerFunctionAnalysis<llvm::ScopedNoAliasAA>();
+        aam.registerFunctionAnalysis<llvm::TypeBasedAA>();
+
+        // Add support for querying global aliasing information when available.
+        // Because the `AAManager` is a function analysis and `GlobalsAA` is a module
+        // analysis, all that the `AAManager` can do is query for any *cached*
+        // results from `GlobalsAA` through a readonly proxy.
+        // aam.registerModuleAnalysis<llvm::GlobalsAA>();
+
+        // Add target-specific alias analyses.
+        if (targetMachine) {
+            targetMachine->registerDefaultAliasAnalyses(aam);
+        }
+        fam.registerPass([aam] { return std::move(aam); });
     }
 }
 
-template <typename T> void DebugModulePassManager::addPostOrderCGSCCPass(T &&P, int stage) {
-    Assert(!m_isFPMOpen && "FunctionPassManager must be committed before adding PostOrderCGSCC passes.");
-    Assert(!m_isLPMOpen && "LoopPassManager must be committed before adding PostOrderCGSCC passes.");
-    // taking number of optimization
-    m_passNumber = (stage == -1) ? (m_passNumber + 1) : stage;
-    if (g->off_stages.find(m_passNumber) == g->off_stages.end()) {
-        // Add PostOrderCGSCC pass to the ModulePassManager directly through adaptor
-        mpm.addPass(llvm::createModuleToPostOrderCGSCCPassAdaptor(std::move(P)));
-        addPassAndDebugPrint(T::name().str(), DebugModulePassManager::Passes::Module);
-    }
-}
+llvm::PreservedAnalyses DebugModulePassManager::run() { return mpm.run(*m, mam); }
 
-template <typename T> void DebugModulePassManager::addFunctionPass(T &&P, int stage) {
-    Assert(m_isFPMOpen && "FunctionPassManager must be initialized before adding function passes");
-    // taking number of optimization
-    m_passNumber = (stage == -1) ? (m_passNumber + 1) : stage;
-    if (g->off_stages.find(m_passNumber) == g->off_stages.end()) {
-        fpmVec.back()->addPass(std::move(P));
-        addPassAndDebugPrint(T::name().str(), DebugModulePassManager::Passes::Function);
-    }
-}
+void DebugModulePassManager::setMemorySSA(bool v) { m_memorySSA = v; }
 
-template <typename T> void DebugModulePassManager::addLoopPass(T &&P, int stage) {
-    Assert(m_isLPMOpen && "LoopPassManager must be initialized before adding function passes");
-    // taking number of optimization
-    m_passNumber = (stage == -1) ? (m_passNumber + 1) : stage;
-    if (g->off_stages.find(m_passNumber) == g->off_stages.end()) {
-        // if not debug stage, add pass to loop pass manager
-        lpmVec.back()->addPass(std::move(P));
-        addPassAndDebugPrint(T::name().str(), DebugModulePassManager::Passes::Loop);
-    }
-}
+void DebugModulePassManager::setBlocksFreq(bool v) { m_blocksFreq = v; }
 
 // Add pass to pass manager and print IR if needed
 void DebugModulePassManager::addPassAndDebugPrint(std::string name, DebugModulePassManager::Passes kind) {
